+++ CMakeLists.txt
@@ -44,12 +44,12 @@ target_sources(${PROJECT_NAME} PRIVATE
 	src/Broadcaster.cpp
 	src/main.cpp
 	src/MediaStreamTrackFactory.cpp
-	src/fake_audio_capture_module.cc
-	src/fake_frame_source.cc
-	src/frame_generator_capturer.cc
-	src/frame_generator.cc
-	src/test_video_capturer.cc
-	src/frame_utils.cc
+	src/fakegenerators/fake_audio_capture_module.cc
+	src/fakegenerators/fake_frame_source.cc
+	src/fakegenerators/frame_generator_capturer.cc
+	src/fakegenerators/frame_generator.cc
+	src/fakegenerators/test_video_capturer.cc
+	src/fakegenerators/frame_utils.cc
 )
 
 # Private (implementation) header files.
