+++ include/fakegenerators/fake_audio_capture_module.h
