+++ include/fakegenerators/fake_frame_source.h
