+++ include/fakegenerators/frame_generator.h
