+++ include/fakegenerators/frame_utils.h
