+++ include/fakegenerators/test_video_capturer.h
