+++ src/fakegenerators/fake_audio_capture_module.cc
@@ -8,7 +8,7 @@
  *  be found in the AUTHORS file in the root of the source tree.
  */
 
-#include "pc/test/fake_audio_capture_module.h"
+#include "fakegenerators/fake_audio_capture_module.h"
 
 #include <string.h>
