+++ src/fakegenerators/frame_generator_capturer.cc
@@ -8,7 +8,7 @@
  *  be found in the AUTHORS file in the root of the source tree.
  */
 
-#include "frame_generator_capturer.h"
+#include "fakegenerators/frame_generator_capturer.h"
 
 #include <algorithm>
 #include <cmath>
